Reuse abstract dropdown editor.
