+++ src/article/shared/LicenseEditor.js
@@ -1,41 +1,12 @@
-import {
-  ValueComponent
-} from '../../kit'
+import DropdownEditor from './DropdownEditor'
 import { LICENSES } from '../ArticleConstants'
 
-export default class LicenseEditor extends ValueComponent {
-  render ($$) {
-    const model = this.props.model
-    const value = model.getValue()
-    let el = $$('div').addClass('sc-license-editor')
-
-    const licenseSelector = $$('select').ref('input').addClass('se-select')
-      .on('click', this._suppressClickPropagation)
-      .on('change', this._setLicense)
-
-    licenseSelector.append(
-      $$('option').append(this.getLabel('select-license'))
-    )
-
-    LICENSES.forEach(l => {
-      const option = $$('option').attr({ value: l.id }).append(l.name)
-      if (l.id === value) option.attr({ selected: 'selected' })
-      licenseSelector.append(option)
-    })
-
-    el.append(licenseSelector)
-
-    return el
+export default class LicenseEditor extends DropdownEditor {
+  _getLabel () {
+    return this.getLabel('select-license')
   }
 
-  _setLicense () {
-    const model = this.props.model
-    const input = this.refs.input
-    const value = input.getValue()
-    model.setValue(value)
-  }
-
-  _suppressClickPropagation (e) {
-    e.stopPropagation()
+  _getValues () {
+    return LICENSES
   }
 }
